feat: stabilize phase2 and enable phase3 relower
Fix phase2 control-flow regressions, widen instruction-length metadata to u16, and restore release tail threading so the phase3 relower can stay always on.

+++ crates/telomere/src/common.rs
@@ -329,7 +329,7 @@ pub struct Global(pub GlobalType, pub Vec<ConstExpr>);
 pub struct Func {
     pub locals: LocalsData,
     pub expr: Vec<Instr>,
-    pub op_lens: Vec<u8>,
+    pub op_lens: Vec<u16>,
 }
 impl Func {
     pub fn local_size(&self) -> usize {
